Fix AAD Kerb storage account service principal search
The default behavior of Get-AzureADServicePrincipal is to return the
first 100 service principals. For tenants with more than 100 service
principals, this command would therefore not always find the service
principal.

Using -Filter instead of Where-Object moves the filtering from the
client to the server, and thereby avoids the limit of only 100 service
principals being returned.

The documentation of Get-AzureADServicePrincipal is also being fixed
to try to avoid this issue in the future.